Update imports to reflect nodes modularisation
