+++ plugins/nodes/src/js/components/__tests__/NodesGridDials-test.js
@@ -11,8 +11,8 @@ const ReactDOM = require('react-dom');
 const TestUtils = require('react-addons-test-utils');
 
 const NodesGridDials = require('../NodesGridDials');
-const ResourcesUtil = require('../../utils/ResourcesUtil');
-const Node = require('../../structs/Node');
+const ResourcesUtil = require('../../../../../../src/js/utils/ResourcesUtil');
+const Node = require('../../../../../../src/js/structs/Node');
 
 var mockHost = {
   id: 'foo',
