+++ plugins/nodes/src/js/components/__tests__/NodesGridView-test.js
@@ -8,10 +8,10 @@ const React = require('react');
 /* eslint-enable no-unused-vars */
 const ReactDOM = require('react-dom');
 
-var CompositeState = require('../../structs/CompositeState');
+var CompositeState = require('../../../../../../src/js/structs/CompositeState');
 var NodesGridView = require('../NodesGridView');
-var MesosStateStore = require('../../stores/MesosStateStore');
-var NodesList = require('../../structs/NodesList');
+var MesosStateStore = require('../../../../../../src/js/stores/MesosStateStore');
+var NodesList = require('../../../../../../src/js/structs/NodesList');
 
 MesosStateStore.addChangeListener = function () {};
