+++ plugins/nodes/src/js/pages/NodesOverview.js
@@ -3,21 +3,21 @@ import React from 'react';
 import {HashLocation, Link, RouteHandler} from 'react-router';
 import {StoreMixin} from 'mesosphere-shared-reactjs';
 
-import AlertPanel from '../components/AlertPanel';
-import CompositeState from '../structs/CompositeState';
-import Config from '../config/Config';
-import EventTypes from '../constants/EventTypes';
-import FilterButtons from '../components/FilterButtons';
-import FilterBar from '../components/FilterBar';
-import FilterByService from '../../../plugins/services/src/js/components/FilterByService';
-import FilterInputText from '../components/FilterInputText';
-import FilterHeadline from '../components/FilterHeadline';
-import Icon from '../components/Icon';
-import InternalStorageMixin from '../mixins/InternalStorageMixin';
-import MesosSummaryStore from '../stores/MesosSummaryStore';
-import ResourceBarChart from '../components/charts/ResourceBarChart';
-import SidebarActions from '../events/SidebarActions';
-import StringUtil from '../utils/StringUtil';
+import AlertPanel from '../../../../../src/js/components/AlertPanel';
+import CompositeState from '../../../../../src/js/structs/CompositeState';
+import Config from '../../../../../src/js/config/Config';
+import EventTypes from '../../../../../src/js/constants/EventTypes';
+import FilterButtons from '../../../../../src/js/components/FilterButtons';
+import FilterBar from '../../../../../src/js/components/FilterBar';
+import FilterByService from '../../../../services/src/js/components/FilterByService';
+import FilterInputText from '../../../../../src/js/components/FilterInputText';
+import FilterHeadline from '../../../../../src/js/components/FilterHeadline';
+import Icon from '../../../../../src/js/components/Icon';
+import InternalStorageMixin from '../../../../../src/js/mixins/InternalStorageMixin';
+import MesosSummaryStore from '../../../../../src/js/stores/MesosSummaryStore';
+import ResourceBarChart from '../../../../../src/js/components/charts/ResourceBarChart';
+import SidebarActions from '../../../../../src/js/events/SidebarActions';
+import StringUtil from '../../../../../src/js/utils/StringUtil';
 
 const NODES_DISPLAY_LIMIT = 300;
 const HEALTH_FILTER_BUTTONS = ['all', 'healthy', 'unhealthy'];
