+++ plugins/nodes/src/js/pages/NodesPage.js
@@ -1,8 +1,8 @@
 import {RouteHandler} from 'react-router';
 import React from 'react';
 
-import Icon from '../components/Icon';
-import Page from '../components/Page';
+import Icon from '../../../../../src/js/components/Icon';
+import Page from '../../../../../src/js/components/Page';
 
 class NodesPage extends React.Component {
   render() {
