+++ plugins/nodes/src/js/pages/__tests__/NodeDetailPage-test.js
@@ -6,7 +6,7 @@ jest.dontMock('../nodes/NodeDetailPage');
 jest.dontMock('../../components/RequestErrorMsg');
 jest.dontMock('../../structs/CompositeState');
 
-const JestUtil = require('../../utils/JestUtil');
+const JestUtil = require('../../../../../../src/js/utils/JestUtil');
 
 JestUtil.unMockStores(['MesosSummaryStore', 'MesosStateStore']);
 /* eslint-disable no-unused-vars */
@@ -15,12 +15,12 @@ const React = require('react');
 const ReactDOM = require('react-dom');
 const TestUtils = require('react-addons-test-utils');
 
-const CompositeState = require('../../structs/CompositeState');
-const MesosStateStore = require('../../stores/MesosStateStore');
-const MesosSummaryActions = require('../../events/MesosSummaryActions');
-const MesosSummaryStore = require('../../stores/MesosSummaryStore');
-const Node = require('../../structs/Node');
-const NodesList = require('../../structs/NodesList');
+const CompositeState = require('../../../../../../src/js/structs/CompositeState');
+const MesosStateStore = require('../../../../../../src/js/stores/MesosStateStore');
+const MesosSummaryActions = require('../../../../../../src/js/events/MesosSummaryActions');
+const MesosSummaryStore = require('../../../../../../src/js/stores/MesosSummaryStore');
+const Node = require('../../../../../../src/js/structs/Node');
+const NodesList = require('../../../../../../src/js/structs/NodesList');
 const NodeDetailPage = require('../nodes/NodeDetailPage');
 
 describe('NodeDetailPage', function () {
