+++ plugins/nodes/src/js/routes/nodes.js
@@ -13,15 +13,15 @@ import NodesGridView from '../components/NodesGridView';
 import NodesOverview from '../pages/NodesOverview';
 import NodesPage from '../pages/NodesPage';
 
-import TaskDetail from '../../../plugins/services/src/js/pages/task-details/TaskDetail';
-import TaskDetailBreadcrumb from '../../../plugins/services/src/js/pages/nodes/breadcrumbs/TaskDetailBreadcrumb';
-import TaskDetailsTab from '../../../plugins/services/src/js/pages/task-details/TaskDetailsTab';
-import TaskFilesTab from '../../../plugins/services/src/js/pages/task-details/TaskFilesTab';
-import TaskFileViewer from '../../../plugins/services/src/js/pages/task-details/TaskFileViewer';
-import UnitsHealthNodeDetail from '../pages/system/UnitsHealthNodeDetail';
-import UnitsHealthDetailBreadcrumb from '../pages/system/breadcrumbs/UnitsHealthDetailBreadcrumb';
-import VolumeDetail from '../../../plugins/services/src/js/components/VolumeDetail';
-import VolumeTable from '../../../plugins/services/src/js/components/VolumeTable';
+import TaskDetail from '../../../../services/src/js/pages/task-details/TaskDetail';
+import TaskDetailBreadcrumb from '../../../../services/src/js/pages/nodes/breadcrumbs/TaskDetailBreadcrumb';
+import TaskDetailsTab from '../../../../services/src/js/pages/task-details/TaskDetailsTab';
+import TaskFilesTab from '../../../../services/src/js/pages/task-details/TaskFilesTab';
+import TaskFileViewer from '../../../../services/src/js/pages/task-details/TaskFileViewer';
+import UnitsHealthNodeDetail from '../../../../../src/js/pages/system/UnitsHealthNodeDetail';
+import UnitsHealthDetailBreadcrumb from '../../../../../src/js/pages/system/breadcrumbs/UnitsHealthDetailBreadcrumb';
+import VolumeDetail from '../../../../services/src/js/components/VolumeDetail';
+import VolumeTable from '../../../../services/src/js/components/VolumeTable';
 
 let nodesRoutes = {
   type: Route,
