+++ plugins/nodes/src/js/stores/__tests__/NodeHealthStore-test.js
@@ -6,13 +6,13 @@ jest.dontMock('../../../../tests/_fixtures/unit-health/nodes.json');
 
 const RequestUtil = require('mesosphere-shared-reactjs').RequestUtil;
 
-const ActionTypes = require('../../constants/ActionTypes');
-const AppDispatcher = require('../../events/AppDispatcher');
-const Config = require('../../config/Config');
-const EventTypes = require('../../constants/EventTypes');
-const NodesList = require('../../structs/NodesList');
+const ActionTypes = require('../../../../../../src/js/constants/ActionTypes');
+const AppDispatcher = require('../../../../../../src/js/events/AppDispatcher');
+const Config = require('../../../../../../src/js/config/Config');
+const EventTypes = require('../../../../../../src/js/constants/EventTypes');
+const NodesList = require('../../../../../../src/js/structs/NodesList');
 const NodeHealthStore = require('../NodeHealthStore');
-const nodesFixture = require('../../../../tests/_fixtures/unit-health/nodes.json');
+const nodesFixture = require('../../../../../../tests/_fixtures/unit-health/nodes.json');
 
 describe('NodeHealthStore', function () {
