+++ src/js/routes/index.js
@@ -6,7 +6,7 @@ import components from './components';
 import dashboard from './dashboard';
 import Index from '../pages/Index';
 import Network from './factories/network';
-import nodes from './nodes';
+import nodes from '../../../plugins/nodes/src/js/routes/nodes';
 import NotFoundPage from '../pages/NotFoundPage';
 import Organization from './factories/organization';
 import settings from './settings';
